feat:add prefetching for spotify and visitcount using react-query
