+++ src/app/layout.tsx
@@ -7,6 +7,9 @@ import Footer from "@/components/layout/Footer";
 import { SpeedInsights } from "@vercel/speed-insights/next"
 import { Analytics } from '@vercel/analytics/react';
 import { Providers } from "@/components/providers/providers";
+import { HydrationBoundary, dehydrate } from '@tanstack/react-query';
+import { prefetchAll } from '@/lib/prefetch';
+import getQueryClient from '@/lib/getQueryClient';
 
 const raleway = Raleway({
   subsets: ["latin"],
@@ -37,25 +40,33 @@ export const metadata: Metadata = {
   },
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const queryClient = getQueryClient();
+  await prefetchAll();
+  const dehydratedState = dehydrate(queryClient);
+
   return (
     <html lang="en" suppressHydrationWarning>
       <body
         suppressHydrationWarning
         className={`${raleway.className} mx-auto flex min-h-screen max-w-3xl flex-col px-8 pb-16 antialiased`}
       >
+
         <Providers>
-          <Header />
-          <main className="grow">
-            {children}
-            <Analytics />
-            <SpeedInsights />
-          </main>
-          <Footer />
+          <HydrationBoundary state={dehydratedState}>
+            <Header />
+            <main className="grow">
+              {children}
+              <Analytics />
+              <SpeedInsights />
+            </main>
+            <Footer />
+          </HydrationBoundary>
+
         </Providers>
       </body>
     </html>
