+++ src/components/providers/providers.tsx
@@ -4,11 +4,13 @@ import { ThemeProvider } from "@/context/theme/theme-provider"
 import { ChatProvider } from "../../context/chatbot/chat-context"
 import { Toaster } from "../ui/toaster"
 import Chatbot from "../chatbot/Chatbot"
-import { useState } from "react"
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import getQueryClient from "@/lib/getQueryClient"
 
 export function Providers({ children }: { children: React.ReactNode }) {
-    const [queryClient] = useState(() => new QueryClient());
+    const queryClient = getQueryClient()
+
+    
 
     return (
         <ThemeProvider
